Fix PostgreSQL provider to support older versions of PostgreSQL.  Fix connection handling.
As a necessary byproduct of supporting older versions of PostgreSQL
where shared_buffer handling is different, it was also necessary to
fix the connection handling.  Previously the provider would check out
as many connections as it wanted and would basically never free them
or return them back to the sql.DB pool.  After this change each query
performs the appropriate Exec and returns the connection back to the
pool.  As a result, all tests pass and only ~2 connections are opened
at a time.  I've set a default max_connections limit in the provider
of 4 but am allowing users to change this as they see fit in the
provider config.

Fixes: #8 #9

This PR pulled in the fix from #10 as it was required to support PostgreSQL 9.2..