Allow non superuser users to drop databases owned by other roles
Previously allowed non superusers to create databases owned by another user but running a destroy from that same config would error as the connected user isn't an owner of the database.

This is mostly important for Postgres instances in AWS RDS because the user created with the instance doesn't have super user permissions.