grant/default privileges: Fix schema verification.
To check if `postgresql_grant` or `postgresql_default_privileges` resources exists,
the provider was first checking if database/schema for which they will be applied exist.

For the schema, we created a specific connection on the right database to check
it it was mistakenly using the wrong connection and was checking if schema exists
on the database defined in provider.

(Basically it was working only for schema `public` which exists by default in all databases)