Fix error in withRolesGranted behavior.
withRolesGranted (which temporarily grants a role to the admin user used by the provider)
raised an error in a specific case.

If we wanted to temporarily grant role `foo` to the admin user
(usually `postgres`, so `GRANT foo TO postgres`)
it was failing if the reverse membership existed
(so if `foo` was a member of `postgres`).

This fix temporarily revokes this membership if needed.