Use lazy connections.
This avoids the provider to try to connect even if no resources needed.
Also it also avoids the provider to try to connect when defining an RDS instance (for example) and some postgres resources in the same state.
(otherwise, at the first creation, provider will try to connect on a server which does not exist yet)