feat: add postgresql_physical_replication_slot
In our usecase at Veepee we use patroni to bootstrap PostgreSQL clusters and terraform to bootstrap users & databases.
For our Multi Datacenters setup we need fine grain replication slots in order to permit to promote a standby-leader
and move replication slots at the same time, to prevent empty slots.

The current postgresql_replication_slot is unable to handle the physical and as it's clearly not exactly the same way to handle both
we created  a custom resource for physical only