Add new db connection method
This will replace the standard "gocloud.dev/postgres" one given that it lacks the
needed visibility over the CertSource Strut.

Further more we would have liked to reuse the URL Parser that gocloud.dev uses,
however given the visibility scopes given to it, it's not possible to access them.
As a replacement, we use the standard url.Parse() method which works well.