Mention SOCKS5 proxy support to usage documentation
As part of this I found out the format of the values in the environment
variables isn't standardized, so there only place we potentially could link to
about the format of the values would be the source code for
`golang.org/x/net/proxy`. I thought that was a bit much, so I've instead
included a simple example instead.