postgresql_database: Reassign objects owners if database owner changes (#458)
Hi,

this is my first Pull Request, so if something is missing or wrong, let
me know.

To resolve
[439](https://github.com/cyrilgdn/terraform-provider-postgresql/issues/439)
I suggest to add an additional parameter to the database resource called
`alter_object_ownership`. If set to true, it will issue a REASSIGN OWNED
BY query on the database when the owner changes, transfer the whole
ownership in the database from the previous owner to the new one.

As far as I can tell this is only possible if the user in the provider
configuration is a superuser or a user which has grants for both the
previous owner role and the new one. So I skipped the test in the
rds-like test suite.

---------

Co-authored-by: Lukas Albani <lukas.albani@deutschebahn.com>
Co-authored-by: Cyril Gaudin <cyril.gaudin@gmail.com>